refactoring, spacing, linting, tweaks.
