+++ .env.development
@@ -1,8 +1,12 @@
 # Use port 4000 if you want to develop against a local api server
-# Port 3080 is meant to be the fake json-server
 # REACT_APP_API_HOST=http://localhost:4000
-#
-# try using https://create-react-app.dev/docs/proxying-api-requests-in-development/
-REACT_APP_API_HOST=https://localhost:3000
+
+# Port 3080 is meant to be the fake json-server
+REACT_APP_API_HOST=http://localhost:3080
+
+# proxies to production
+# see https://create-react-app.dev/docs/proxying-api-requests-in-development/
+# REACT_APP_API_HOST=https://localhost:3000
+
 REACT_APP_STRIPE_API_KEY=pk_test_6tMDkFKTz4N0wIFQZHuzOUyW
