+++ src/Pages/Apps/Databases/DatabaseCol.tsx
@@ -4,29 +4,29 @@ import classnames from 'classnames'
 import compose from 'recompose/compose'
 import { createStyles, withStyles, WithStyles } from '@material-ui/core/styles'
 import { Grid } from '@material-ui/core'
-import { Alert } from 'antd'
-import moment from 'moment'
 import { StandardDatabase, FreeDatabase } from '../../../api/databases'
+import DbLimitedAt from './DbLimitedAt'
+import DbField from './DbField'
 
 const styles = createStyles({
   divider: {
     width: '100%',
-    borderTop: '4px solid #b7b7b7',
+    borderTop: '1px solid rgba(0,0,0,0.1)',
     height: '0px',
     padding: 0,
     margin: 0,
     marginBottom: '1.33em'
   },
   dividingBorder: {
-    borderRight: '2px solid rgba(224, 224, 224, 1)',
+    borderRight: '1px solid rgba(0,0,0,0.1)',
     minHeight: '100%'
   },
   container: {
-    marginBottom: '1.33em',
-    marginTop: '1.33em'
+    // marginBottom: '1.33em',
+    // marginTop: '1.33em'
   },
   dbLimitedAt: {
-    padding: '0 1.33em 1.33em'
+    padding: '0 20px 20px 0'
   }
 })
 
@@ -48,11 +48,6 @@ const Component: FunctionComponent<DatabaseCol & EnhancedProps> = ({
         dividingBorder === true ? classnames(classes.dividingBorder) : ''
       } ${classnames(classes.container)}`}
     >
-      {database.tier === 'FREE' && database.limitedAt !== null ? (
-        <div className={classnames(classes.dbLimitedAt)}>
-          <DbLimitedAt limitedAt={database.limitedAt} />
-        </div>
-      ) : null}
       <div className={classnames(classes.container)}>
         <Grid container spacing={24}>
           <Grid item xs={3}>
@@ -63,6 +58,12 @@ const Component: FunctionComponent<DatabaseCol & EnhancedProps> = ({
             &nbsp;
           </Grid>
         </Grid>
+        {database.tier === 'FREE' && database.limitedAt !== null ? (
+          <div className={classnames(classes.dbLimitedAt)}>
+            <DbLimitedAt limitedAt={database.limitedAt} />
+          </div>
+        ) : null}
+
         <DbField leftSide="ID" rightSide={database.id} />
         <DbField leftSide="Host" rightSide={database.host} />
         <DbField leftSide="Port" rightSide={database.port} />
@@ -77,90 +78,12 @@ const Component: FunctionComponent<DatabaseCol & EnhancedProps> = ({
             <DbField leftSide="Size" rightSide={database.size} />
           </>
         ) : (
-          <DbFieldLimitedAt limitedAt={database.limitedAt} />
+          <span />
         )}
 
         <DbField leftSide="URL" rightSide={database.url} />
       </div>
     </div>
-  )
-}
-
-const DbField: FunctionComponent<{
-  leftSide: string | number
-  rightSide: string | number
-}> = ({ leftSide, rightSide }) => {
-  if (
-    typeof rightSide === 'undefined' ||
-    rightSide === null ||
-    (typeof rightSide === 'string' && rightSide.trim() === '')
-  )
-    return null
-  return (
-    <Grid container spacing={24}>
-      <Grid item xs={2} style={{ fontWeight: 'bold' }}>
-        {leftSide}
-      </Grid>
-      <Grid item xs={9}>
-        {rightSide}
-      </Grid>
-    </Grid>
-  )
-}
-
-const DbFieldLimitedAt: FunctionComponent<{
-  limitedAt: Date
-}> = ({ limitedAt }) => {
-  if (typeof limitedAt === 'undefined' || limitedAt === null) return null
-  const limitedAtMoment = moment(limitedAt)
-  return (
-    <Grid container spacing={24}>
-      <Grid item xs={2} style={{ fontWeight: 'bold' }}>
-        Limited At
-      </Grid>
-      <Grid item xs={9}>
-        {limitedAtMoment.format('h:mma')} on{' '}
-        {limitedAtMoment.format('MM/DD/YYYY')}
-      </Grid>
-    </Grid>
-  )
-}
-
-const DbFieldTextArea: FunctionComponent<{
-  leftSide: string | number
-  rightSide: string | number
-}> = ({ leftSide, rightSide }) => {
-  if (typeof rightSide === 'undefined' || rightSide === null) return null
-  return (
-    <Grid container spacing={24}>
-      <Grid item xs={2} style={{ fontWeight: 'bold' }}>
-        {leftSide}
-      </Grid>
-      <Grid item xs={9}>
-        <textarea style={{ width: '100%', resize: 'none' }}>
-          {rightSide}
-        </textarea>
-      </Grid>
-    </Grid>
-  )
-}
-
-const DbLimitedAt: FunctionComponent<{
-  limitedAt: Date
-}> = ({ limitedAt }) => {
-  if (typeof limitedAt === 'undefined' || limitedAt === null) return null
-  const limitedAtMoment = moment(limitedAt)
-  return (
-    <Alert
-      message="Warning"
-      description={`This database was limited at ${limitedAtMoment.format(
-        'h:mma'
-      )} on ${limitedAtMoment.format(
-        'MM/DD/YYYY'
-      )} because it exceeded 10.000 rows`}
-      type="error"
-      showIcon
-    />
   )
 }
